fixing reference errors
