+++ packages/fast-color-explorer/app/colors.ts
@@ -1,0 +1,21 @@
+export enum AccentColors {
+    black = "#000000",
+    white = "#FFFFFF",
+    blue = "#0078D4",
+    green = "#107C10",
+    purple = "#5C2D91",
+    orange = "#D83B01",
+    yellow = "#F2C812",
+}
+
+export const neutralColors: string[] = [
+    "#808080",
+    "#73818C",
+    "#718E71",
+    "#7F738C",
+    "#8C7A73",
+    "#0078D4",
+    "#107C10",
+    "#5C2D91",
+    "#D83B01",
+];
